requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
bcrypt==4.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.110.2
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.37.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.29.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
